refactor(parse-java): decompose extract.rs into sub-modules, add 14 edge-case tests
Phase 3 - Module decomposition:
- Split 1446-line extract.rs into 7 focused modules:
  constant.rs (75 lines) - SQL detection constants
  string_decode.rs (109 lines) - Java string literal decoding
  heuristics.rs (135 lines) - SQL detection and placeholder conversion
  annotation.rs (142 lines) - annotation SQL extraction
  method_call.rs (97 lines) - method call SQL extraction
  variable.rs (608 lines) - variable/SB declaration and concatenation
  extract.rs (310 lines) - entry point and core visitor dispatch

Phase 4 - Test coverage hardening:
- 14 new edge-case tests covering: colon placeholders with underscores,
  NamedNativeQuery query key, nativeQuery=false JPQL, @SqlQuery,
  executeQuery/executeUpdate, ambiguous method filtering, interface
  declarations, text block indentation, escape sequences, empty files,
  syntax errors, field constant placeholders

All 1160 tests pass (63 java tests).
